ajuste no design e adição de logs

+++ src/main/java/br/com/caelum/alura/service/DispositivoService.java
@@ -3,6 +3,7 @@ package br.com.caelum.alura.service;
 import java.io.IOException;
 import java.util.List;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
@@ -16,7 +17,8 @@ import br.com.caelum.alura.repository.DispositivoRepository;
 @Service
 public class DispositivoService {
 
-	private DispositivoRepository dispositivoRepository;
+	private final DispositivoRepository dispositivoRepository;
+	private static final Logger LOGGER = Logger.getLogger(AlunoService.class);
 
 	@Autowired
 	public DispositivoService(DispositivoRepository dispositivoRepository) {
@@ -25,23 +27,40 @@ public class DispositivoService {
 
 	public void salva(Dispositivo dispositivo) {
 		dispositivoRepository.save(dispositivo);
+		LOGGER.info("dispositivo salvo " + dispositivo.getToken());
 	}
 
 	@Async
 	public void enviaNotificacao(List<Aluno> alunos) {
+		logaAlunos("enviando alunos", alunos);
 
 		List<Dispositivo> dispositivos = (List<Dispositivo>) dispositivoRepository.findAll();
 		try {
-			Thread.sleep(1000L);
 			FirebaseSender firebaseSender = new FirebaseSender(this);
 			firebaseSender.envia(dispositivos, new AlunoSync(alunos));
-		} catch (IOException | InterruptedException e) {
+
+		} catch (IOException e) {
 			e.printStackTrace();
 		}
+
+		logaAlunos("alunos enviados com sucesso", alunos);
 	}
 
 	public void deleta(List<Dispositivo> invalidos) {
-		dispositivoRepository.delete(invalidos);
+		if (!invalidos.isEmpty()) {
+			dispositivoRepository.delete(invalidos);
+			logaDispositivos("dispositivos excluídos", invalidos);
+		}
+	}
+
+	private void logaAlunos(String mensagem, List<Aluno> alunos) {
+		LOGGER.info(mensagem);
+		alunos.forEach(aluno -> LOGGER.info(aluno.getId()));
+	}
+
+	private void logaDispositivos(String mensagem, List<Dispositivo> invalidos) {
+		LOGGER.info(mensagem);
+		invalidos.forEach(dispositivo -> LOGGER.info(dispositivo.getToken()));
 	}
 
 }
